Codebase refinement: add DB indexes, fix sync bugs, decompose screens
Quality pass before Phase 3 features. Four workstreams:

WS1 - Database indexes + schema v3 migration:
  - Add 13 indexes on high-query columns across all major tables
  - Add consecutiveFailures, lastFailureTime, isSyncing columns to
    BankConnections for circuit breaker persistence and sync locking
  - @TableIndex annotations ensure fresh installs get indexes via createAll()

WS2 - Shared UI utilities (new files only):
  - showErrorSnackbar/showSuccessSnackbar helpers
  - showDeleteConfirmation reusable dialog

WS3 - Screen decomposition (5 screens → 13 extracted widgets):
  - Dashboard 465→78 lines (5 card widgets extracted)
  - Transactions 600→188 lines (filter sheet + tile extracted)
  - CSV Import 716→399 lines (mapping + preview steps extracted)
  - Goals 466→399 lines (appearance section extracted)
  - Recurring 469→435 lines (account/category section extracted)

WS4 - SimpleFIN sync fixes:
  - Persist circuit breaker state in DB (survives restarts + isolates)
  - Atomic sync lock via UPDATE WHERE is_syncing=0 (cross-isolate safe)
  - Dedicated 60s-timeout Dio client for transaction pulls
  - Background sync only attempts ConnectionStatus.connected
  - Stale lock recovery on app startup and background callback start
