+++ lib/main.dart
@@ -8,6 +8,7 @@ import 'package:workmanager/workmanager.dart';
 import 'app.dart';
 import 'core/di/providers.dart';
 import 'data/local/database/app_database.dart';
+import 'data/repositories/bank_connection_repository.dart';
 import 'data/repositories/category_repository.dart';
 import 'domain/usecases/categories/category_seeder.dart';
 import 'domain/usecases/sync/background_sync_callback.dart';
@@ -30,6 +31,14 @@ Future<void> main() async {
     await seeder.seedIfEmpty();
   } catch (e) {
     if (kDebugMode) debugPrint('Category seeding failed: $e');
+  }
+
+  // Reset stale sync locks from any previous crash
+  try {
+    final connectionRepo = BankConnectionRepository(database);
+    await connectionRepo.resetAllSyncLocks();
+  } catch (e) {
+    if (kDebugMode) debugPrint('Sync lock reset failed: $e');
   }
 
   // TODO: Initialize Sentry
