+++ lib/presentation/features/import/csv_import_screen.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../core/di/providers.dart';
-import '../../../core/extensions/money_extensions.dart';
 import '../../../core/theme/app_theme.dart';
 import '../../../domain/usecases/import/csv_import_service.dart';
-import '../accounts/accounts_providers.dart';
+import 'widgets/column_mapping_step.dart';
+import 'widgets/preview_step.dart';
 
 /// Multi-step CSV import flow.
 class CsvImportScreen extends ConsumerStatefulWidget {
@@ -233,13 +233,49 @@ class _CsvImportScreenState extends ConsumerState<CsvImportScreen> {
             title: const Text('Map Columns'),
             isActive: _currentStep >= 1,
             state: _currentStep > 1 ? StepState.complete : StepState.indexed,
-            content: _buildMappingStep(),
+            content: ColumnMappingStep(
+              previewRows: _previewRows,
+              headers: _headers,
+              hasHeader: _hasHeader,
+              dateColumn: _dateColumn,
+              dateFormat: _dateFormat,
+              amountColumn: _amountColumn,
+              payeeColumn: _payeeColumn,
+              categoryColumn: _categoryColumn,
+              notesColumn: _notesColumn,
+              negativeIsExpense: _negativeIsExpense,
+              isParsing: _isParsing,
+              onDateColumnChanged: (v) =>
+                  setState(() => _dateColumn = v ?? 0),
+              onDateFormatChanged: (v) =>
+                  setState(() => _dateFormat = v ?? 'MM/dd/yyyy'),
+              onAmountColumnChanged: (v) =>
+                  setState(() => _amountColumn = v ?? 1),
+              onPayeeColumnChanged: (v) =>
+                  setState(() => _payeeColumn = v ?? 2),
+              onCategoryColumnChanged: (v) =>
+                  setState(() => _categoryColumn = v),
+              onNotesColumnChanged: (v) =>
+                  setState(() => _notesColumn = v),
+              onNegativeIsExpenseChanged: (v) =>
+                  setState(() => _negativeIsExpense = v ?? true),
+              onParsePressed: _parseWithConfig,
+              onBackPressed: () => setState(() => _currentStep = 0),
+            ),
           ),
           Step(
             title: const Text('Preview & Import'),
             isActive: _currentStep >= 2,
             state: _currentStep > 2 ? StepState.complete : StepState.indexed,
-            content: _buildPreviewStep(),
+            content: PreviewStep(
+              preview: _preview,
+              selectedAccountId: _selectedAccountId,
+              isImporting: _isImporting,
+              onAccountChanged: (v) =>
+                  setState(() => _selectedAccountId = v),
+              onImportPressed: _executeImport,
+              onBackPressed: () => setState(() => _currentStep = 1),
+            ),
           ),
           Step(
             title: const Text('Results'),
@@ -288,359 +324,6 @@ class _CsvImportScreenState extends ConsumerState<CsvImportScreen> {
           onPressed: _loadFile,
           icon: const Icon(Icons.upload_file),
           label: const Text('Load File'),
-        ),
-      ],
-    );
-  }
-
-  // ---------------------------------------------------------------------------
-  // Step 2: Column Mapping
-  // ---------------------------------------------------------------------------
-
-  Widget _buildMappingStep() {
-    if (_previewRows.isEmpty) {
-      return const Text('No data loaded. Go back and select a file.');
-    }
-
-    final dataRows = _hasHeader ? _previewRows.skip(1).toList() : _previewRows;
-    final colCount = _previewRows.first.length;
-    final columnOptions = List.generate(
-      colCount,
-      (i) => DropdownMenuItem(
-        value: i,
-        child: Text(
-          _hasHeader && i < _headers.length
-              ? 'Col $i: ${_headers[i]}'
-              : 'Column $i',
-        ),
-      ),
-    );
-
-    final optionalColumnOptions = [
-      const DropdownMenuItem<int?>(value: null, child: Text('None')),
-      ...columnOptions.map((item) => DropdownMenuItem<int?>(
-            value: item.value,
-            child: item.child,
-          )),
-    ];
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        // CSV preview table
-        const Text('File Preview:', style: TextStyle(fontWeight: FontWeight.bold)),
-        const SizedBox(height: 8),
-        SingleChildScrollView(
-          scrollDirection: Axis.horizontal,
-          child: DataTable(
-            headingRowHeight: 36,
-            dataRowMinHeight: 32,
-            dataRowMaxHeight: 32,
-            columnSpacing: 16,
-            columns: List.generate(
-              colCount,
-              (i) => DataColumn(
-                label: Text(
-                  _hasHeader && i < _headers.length
-                      ? _headers[i]
-                      : 'Col $i',
-                  style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 12),
-                ),
-              ),
-            ),
-            rows: dataRows.take(3).map((row) {
-              return DataRow(
-                cells: List.generate(
-                  colCount,
-                  (i) => DataCell(Text(
-                    i < row.length ? row[i] : '',
-                    style: const TextStyle(fontSize: 12),
-                  )),
-                ),
-              );
-            }).toList(),
-          ),
-        ),
-
-        const SizedBox(height: 20),
-        const Text('Column Mapping:', style: TextStyle(fontWeight: FontWeight.bold)),
-        const SizedBox(height: 8),
-
-        // Date column
-        DropdownButtonFormField<int>(
-          initialValue: _dateColumn,
-          decoration: const InputDecoration(
-            labelText: 'Date Column *',
-            filled: true,
-            isDense: true,
-          ),
-          items: columnOptions,
-          onChanged: (v) => setState(() => _dateColumn = v ?? 0),
-        ),
-        const SizedBox(height: 12),
-
-        // Date format
-        DropdownButtonFormField<String>(
-          initialValue: _dateFormat,
-          decoration: const InputDecoration(
-            labelText: 'Date Format *',
-            filled: true,
-            isDense: true,
-          ),
-          items: dateFormats
-              .map((f) => DropdownMenuItem(value: f.$1, child: Text(f.$2)))
-              .toList(),
-          onChanged: (v) => setState(() => _dateFormat = v ?? 'MM/dd/yyyy'),
-        ),
-        const SizedBox(height: 12),
-
-        // Amount column
-        DropdownButtonFormField<int>(
-          initialValue: _amountColumn,
-          decoration: const InputDecoration(
-            labelText: 'Amount Column *',
-            filled: true,
-            isDense: true,
-          ),
-          items: columnOptions,
-          onChanged: (v) => setState(() => _amountColumn = v ?? 1),
-        ),
-        const SizedBox(height: 12),
-
-        // Payee column
-        DropdownButtonFormField<int>(
-          initialValue: _payeeColumn,
-          decoration: const InputDecoration(
-            labelText: 'Payee Column *',
-            filled: true,
-            isDense: true,
-          ),
-          items: columnOptions,
-          onChanged: (v) => setState(() => _payeeColumn = v ?? 2),
-        ),
-        const SizedBox(height: 12),
-
-        // Category column (optional)
-        DropdownButtonFormField<int?>(
-          initialValue: _categoryColumn,
-          decoration: const InputDecoration(
-            labelText: 'Category Column',
-            filled: true,
-            isDense: true,
-          ),
-          items: optionalColumnOptions,
-          onChanged: (v) => setState(() => _categoryColumn = v),
-        ),
-        const SizedBox(height: 12),
-
-        // Notes column (optional)
-        DropdownButtonFormField<int?>(
-          initialValue: _notesColumn,
-          decoration: const InputDecoration(
-            labelText: 'Notes Column',
-            filled: true,
-            isDense: true,
-          ),
-          items: optionalColumnOptions,
-          onChanged: (v) => setState(() => _notesColumn = v),
-        ),
-        const SizedBox(height: 12),
-
-        // Sign convention
-        CheckboxListTile(
-          title: const Text('Negative amounts are expenses'),
-          subtitle: const Text('Uncheck if your bank uses positive for expenses'),
-          value: _negativeIsExpense,
-          onChanged: (v) => setState(() => _negativeIsExpense = v ?? true),
-          contentPadding: EdgeInsets.zero,
-        ),
-        const SizedBox(height: 16),
-
-        FilledButton(
-          onPressed: _isParsing ? null : _parseWithConfig,
-          child: _isParsing
-              ? const SizedBox(
-                  width: 20,
-                  height: 20,
-                  child: CircularProgressIndicator(strokeWidth: 2),
-                )
-              : const Text('Parse & Preview'),
-        ),
-
-        const SizedBox(height: 8),
-        OutlinedButton(
-          onPressed: () => setState(() => _currentStep = 0),
-          child: const Text('Back'),
-        ),
-      ],
-    );
-  }
-
-  // ---------------------------------------------------------------------------
-  // Step 3: Preview & Import
-  // ---------------------------------------------------------------------------
-
-  Widget _buildPreviewStep() {
-    final preview = _preview;
-    if (preview == null) {
-      return const Text('No data parsed. Go back and configure column mapping.');
-    }
-
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final accountsAsync = ref.watch(accountsProvider);
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        // Summary card
-        Card(
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text('File: ${preview.fileName}',
-                    style: theme.textTheme.titleSmall),
-                const SizedBox(height: 8),
-                Text('Total rows: ${preview.totalRows}'),
-                Text('Valid transactions: ${preview.transactions.length}'),
-                Text('Duplicates: ${preview.duplicateCount}'),
-                if (preview.errors.isNotEmpty)
-                  Text(
-                    'Errors: ${preview.errors.length}',
-                    style: TextStyle(color: colorScheme.error),
-                  ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(height: 16),
-
-        // Errors (if any)
-        if (preview.errors.isNotEmpty) ...[
-          Text('Errors:', style: TextStyle(
-            fontWeight: FontWeight.bold,
-            color: colorScheme.error,
-          )),
-          const SizedBox(height: 4),
-          ...preview.errors.take(5).map((e) => Padding(
-                padding: const EdgeInsets.only(bottom: 2),
-                child: Text(
-                  'Row ${e.rowNumber}: ${e.message}',
-                  style: TextStyle(fontSize: 12, color: colorScheme.error),
-                ),
-              )),
-          if (preview.errors.length > 5)
-            Text(
-              '... and ${preview.errors.length - 5} more errors',
-              style: TextStyle(fontSize: 12, color: colorScheme.onSurfaceVariant),
-            ),
-          const SizedBox(height: 16),
-        ],
-
-        // Transaction preview list
-        Text('Preview (first 20):', style: theme.textTheme.titleSmall),
-        const SizedBox(height: 8),
-        ...preview.transactions.take(20).map((t) => ListTile(
-              dense: true,
-              contentPadding: EdgeInsets.zero,
-              leading: t.isDuplicate
-                  ? Icon(Icons.content_copy, color: colorScheme.onSurfaceVariant, size: 20)
-                  : Icon(
-                      t.amountCents < 0 ? Icons.arrow_upward : Icons.arrow_downward,
-                      color: t.amountCents < 0
-                          ? theme.finance.expense
-                          : theme.finance.income,
-                      size: 20,
-                    ),
-              title: Text(t.payee, style: const TextStyle(fontSize: 14)),
-              subtitle: Text(
-                '${t.date.toMediumDate()}${t.isDuplicate ? " (duplicate)" : ""}',
-                style: TextStyle(
-                  fontSize: 12,
-                  color: t.isDuplicate
-                      ? colorScheme.onSurfaceVariant
-                      : null,
-                ),
-              ),
-              trailing: Text(
-                t.amountCents.toCurrency(),
-                style: TextStyle(
-                  fontWeight: FontWeight.w500,
-                  color: t.isDuplicate
-                      ? colorScheme.onSurfaceVariant
-                      : t.amountCents < 0
-                          ? theme.finance.expense
-                          : theme.finance.income,
-                ),
-              ),
-            )),
-
-        const SizedBox(height: 20),
-
-        // Account selector
-        accountsAsync.when(
-          loading: () => const LinearProgressIndicator(),
-          error: (_, _) => Text('Error loading accounts',
-              style: TextStyle(color: colorScheme.error)),
-          data: (accounts) {
-            if (accounts.isEmpty) {
-              return Card(
-                color: colorScheme.errorContainer,
-                child: const Padding(
-                  padding: EdgeInsets.all(16),
-                  child: Text('No accounts found. Create an account first.'),
-                ),
-              );
-            }
-
-            _selectedAccountId ??= accounts.first.id;
-
-            return DropdownButtonFormField<String>(
-              initialValue: _selectedAccountId,
-              decoration: const InputDecoration(
-                labelText: 'Import to Account *',
-                filled: true,
-              ),
-              items: accounts
-                  .map((a) => DropdownMenuItem(
-                        value: a.id,
-                        child: Text(a.name),
-                      ))
-                  .toList(),
-              onChanged: (v) => setState(() => _selectedAccountId = v),
-            );
-          },
-        ),
-        const SizedBox(height: 16),
-
-        if (preview.validCount > 0)
-          FilledButton(
-            onPressed: _isImporting ? null : _executeImport,
-            child: _isImporting
-                ? const SizedBox(
-                    width: 20,
-                    height: 20,
-                    child: CircularProgressIndicator(strokeWidth: 2),
-                  )
-                : Text('Import ${preview.validCount} Transactions'),
-          ),
-
-        if (preview.validCount == 0)
-          Card(
-            color: colorScheme.errorContainer,
-            child: const Padding(
-              padding: EdgeInsets.all(16),
-              child: Text('No new transactions to import. All rows are either '
-                  'duplicates or have errors.'),
-            ),
-          ),
-
-        const SizedBox(height: 8),
-        OutlinedButton(
-          onPressed: () => setState(() => _currentStep = 1),
-          child: const Text('Back'),
         ),
       ],
     );
