+++ lib/presentation/shared/utils/snackbar_helpers.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+/// Show an error snackbar with the app's error styling.
+void showErrorSnackbar(BuildContext context, String message) {
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(
+      content: Text(message),
+      backgroundColor: Theme.of(context).colorScheme.error,
+    ),
+  );
+}
+
+/// Show a success snackbar.
+void showSuccessSnackbar(BuildContext context, String message) {
+  ScaffoldMessenger.of(context).showSnackBar(
+    SnackBar(content: Text(message)),
+  );
+}
